game now successfully runs past the first econ tick without crashing. Mine processor is now a full class.

+++ Pulsar4X/Pulsar4X.ECSLib/Processors/ShipAndColonyInfoProcessor.cs
@@ -45,7 +45,7 @@ namespace Pulsar4X.ECSLib
 
     public class TonnageAndHTKRecalc : IRecalcProcessor
     {
-        public int ProcessPriority { get; set; } = 0;
+        public byte ProcessPriority { get; set; } = 0;
         public void RecalcEntity(Entity entity)
         {
             ShipAndColonyInfoProcessor.ReCalculateShipTonnaageAndHTK(entity);
